Move actions out of lib directly into /src
